Move `BoardValue` into `crate::model::board_value`
Deserialize `BoardValue` from a string representation and use the
`BoardValue` representation everywhere rather than only translating it
in the `BoardCell` component.
