+++ ui-rs/crates/mines_mogwai/src/components/cell.rs
@@ -1,4 +1,4 @@
-use crate::model::{CellInteract, CellInteractKind, CellUpdate};
+use crate::model::{BoardValue, CellInteract, CellInteractKind, CellUpdate};
 use mogwai::prelude::*;
 
 pub struct BoardCell {
@@ -12,7 +12,7 @@ impl BoardCell {
     pub fn gizmo(
         column: usize,
         row: usize,
-        initial_value: String,
+        initial_value: BoardValue,
         tx: &Transmitter<CellInteract>,
         rx: &Receiver<CellUpdate>,
     ) -> Gizmo<Self> {
@@ -66,7 +66,7 @@ impl Component for BoardCell {
         tx.send(&CellUpdate::Single {
             column: self.column,
             row: self.row,
-            value: board_value.to_string(),
+            value: board_value,
         });
         // Send the `CellInteract` out it (may) eventually result in receiving
         // a `ViewMessage`
@@ -82,16 +82,12 @@ impl Component for BoardCell {
         let col = self.column;
         let row = self.row;
         let rx_value: Receiver<BoardValue> = rx.branch_filter_map(move |update| match update {
-            CellUpdate::All { cells } => cells
-                .get(row)
-                .map(|r| r.get(col))
-                .flatten()
-                .map(|s| s.into()),
+            CellUpdate::All { cells } => cells.get(row).map(|r| r.get(col)).flatten().map(|s| *s),
             CellUpdate::Single {
                 row: y,
                 column: x,
                 value,
-            } if *x == col && *y == row => Some(value.into()),
+            } if *x == col && *y == row => Some(*value),
             _ => None,
         });
         let rx_text = rx_value.branch_map(|update| update.to_string());
@@ -102,51 +98,6 @@ impl Component for BoardCell {
                 // Cells initialize to empty but may update if revealed or clicked
                 {(self.current_display.to_string(), rx_text)}
             </td>
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug)]
-pub enum BoardValue {
-    Closed,
-    Flag,
-    Mine,
-    Open(usize),
-    Pending,
-}
-
-impl std::fmt::Display for BoardValue {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            BoardValue::Closed => f.write_str(" "),
-            BoardValue::Flag => f.write_str("F"),
-            BoardValue::Mine => f.write_str("M"),
-            BoardValue::Open(count) => f.write_fmt(format_args!("{}", count)),
-            BoardValue::Pending => f.write_str("*"),
-        }
-    }
-}
-
-impl<T> From<T> for BoardValue
-where
-    T: AsRef<str>,
-{
-    fn from(value: T) -> Self {
-        let current = value.as_ref();
-        if current == "" || current == " " {
-            BoardValue::Closed
-        } else if current == "M" {
-            BoardValue::Mine
-        } else if current == "F" {
-            BoardValue::Flag
-        } else if current == "*" {
-            BoardValue::Pending
-        } else {
-            use std::str::FromStr;
-            match usize::from_str(&current) {
-                Ok(v) => BoardValue::Open(v),
-                _ => BoardValue::Closed,
-            }
         }
     }
 }
