+++ ui-rs/crates/mines_mogwai/src/model.rs
@@ -1,5 +1,7 @@
+mod board_value;
 mod cell_interact;
 mod cell_update;
 
+pub use board_value::BoardValue;
 pub use cell_interact::{CellInteract, CellInteractKind};
 pub use cell_update::CellUpdate;
